Updated 'DatabaseConnection' to cease program in event of failure to connect.
Added 'Distinct' to store any distinct table fields for use in querying in loops.

Updated 'CSVExport' to add a function to convert objects into String arrays for use in outputting to CSV.

Added features in 'Application' to take user input and output country reports (to be later moved into a separate class), with a function to run a query and create a list of 'CountryReport' objects

Updated 'toString' in 'CountryReport' to tidy up CLI output.